Split out unpackData into separate module
